+++ tests/Functional/Model/Mapping10Document.php
@@ -14,8 +14,12 @@ namespace Sulu\Component\DocumentManager\Tests\Functional\Model;
 class Mapping10Document extends Mapping5Document
 {
     public $six;
+
     public $seven;
+
     public $eight;
+
     public $nine;
+
     public $ten;
 }