+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG for Sulu Document Manager
 ===================================
 
+* 0.10.1 (2018-03-19)
+    * HOTFIX      #119 Refactored version-subscriber to use uuid instead of paths
+    * HOTFIX      #118 Set path and node-name after renaming node
+    * HOTFIX      #117 Execute rename in flush-event to avoid ItemNotFoundException
+
 * 0.10.0 (2017-06-28)
     * FEATURE     #116 Added set-default-author to metadata